avcodec/mediacodec: Add support of dynamic bitrate
MediaCodec supports parameter "video-bitrate" to change the bitrate
on fly. This commit adds capability to use it.

It adds option -bitrate_ctrl_socket to the encoder which makes
the encoder to create an UNIX socket and listen for messages
to change the bitrate.

An example of ffmpeg execution:

    ffmpeg -f flv -i rtmp://0.0.0.0:1935/live/myStream -c:v hevc_mediacodec -bitrate_ctrl_socket /run/bitrate.sock -b:v 8M -f flv rtmp://127.0.0.1:1935/live/reEncoded

An example of changing the bitrate to 1000 Kbps:

    printf '%016X' 1000000 | xxd -r -p | socat -u STDIN UNIX:/run/bitrate.sock

Signed-off-by: Dmitrii Okunev <xaionaro@gmail.com>